requirements.txt
feedparser==6.0.11
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
